change the function used to resolve the glob based on whether the cwd is forced.
when the current working directory is set and thus passed to chokidar
the paths on files emmitted from chokidar are relative.

So in order for anymatch to be able to discern which glob was matched
the globs array should store globs as passed rather than resolving
absolutely. `path.normalize` leaves the glob relative while resolving
any dot parts.

When the cwd is not set the previous approach of using `path.resolve` to
complete relative path to absolute ones works well.