Fix "add" API method
gulp-watch exposes the watchers "add" method, but adding patterns to
watch via this method does not work.

That's because on add/change event, the file path of the added/changed file is
only looked up in the pattern list that was directly passed to gulp-watch. Since
patterns that are added later to the watcher are not in that list, an
error is thrown.

This change provides a custom "add" method implementation which adds the
passed patterns to the internal pattern list.